policy: delete policy with user defined labels on an update
The user installs a CNP with 2 rules defined on which 1 of the
rules contain a user defined label. When doing an CNP update where the
rule with the user defined label no longer exists, that rule would never
be deleted by Cilium because there was no reference to those
user-defined labels.

This commit introduces a new functionality to the PolicyAdd which allows
rules with a specifc set of labels to be deleted during a PolicyAdd.
This should ideally be used during a PolicyAdd to perform an atomic
operation during an update of the CNP.

Signed-off-by: André Martins <andre@cilium.io>