proxylib: Enable parser tests in their own package
Moving most of the functionality not related to the shared library
interface to the proxylib package enables testing in the parser
packages. The actual test parsers are still tested in the main package
to exercise the shared library interface functions.

This commit moves the Cassandra unit tests to the cassandra package
using the new testing tools in the proxylib package. The tests are
somewhat more readable doe to reduced boilerplate.

Accesslog client is simplified now that the path of an existing client
never changes.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>